+++ backend/backend/settings.py
@@ -91,8 +91,8 @@ CORS_ALLOWED_ORIGINS = [
 
 CORS_ALLOW_CREDENTIALS = True
 CORS_ORIGIN_ALLOW_ALL = True
-# SESSION_COOKIE_SAMESITE = None
-# SESSION_COOKIE_SECURE = False
+SESSION_COOKIE_SAMESITE = None
+SESSION_COOKIE_SECURE = True
 # SESSION_COOKIE_SAMESITE = 'Lax'
 
 REST_FRAMEWORK = {
